wikitest: Move 'relpath' to sympy/core/compatibility.py
